Migrate from swaync to dunst
All integrations are in place, including waybar, but are a little bit
limited since some features are missing. I was fed up with latest
updates in swaync just breaking all my stylesheet and have to desire to
maintain configurations for such pieces of software.

The structure of repo is also updated a bit to store the scripts in
corresponding modules instead of just global bin/ dir.